Changed ids to be int64

+++ digtextsimilaritysearch/indexer/faiss_indexer.py
@@ -8,10 +8,10 @@ class FaissIndexer(object):
         self.faiss_index = faiss.read_index(path_to_index_file) \
             if path_to_index_file \
             else faiss.IndexFlatL2(512)
-        # else faiss.IndexIDMap(faiss.IndexFlatL2(512))
+            # else faiss.IndexIDMap(faiss.IndexFlatL2(512))
 
     def index_embeddings(self, embeddings) -> List[int]:
-        ids = np.arange(len(embeddings)) + self.faiss_index.ntotal
+        ids = np.arange(len(embeddings)).astype('int64') + self.faiss_index.ntotal
         self.faiss_index.add(embeddings)
         return ids
